sql/sem/tree: move internalCompare impls to compare.go
Release note: None
