+++ pkg/sql/sem/tree/compare.go
@@ -14,7 +14,13 @@
 
 package tree
 
-import "fmt"
+import (
+	"bytes"
+	"fmt"
+	"math"
+
+	"github.com/cockroachdb/apd"
+)
 
 // This file implements comparisons between tuples.
 //
@@ -138,4 +144,386 @@ func doCompare(ctx *EvalContext, orderedNULLs bool, a, b Datum) int {
 	}
 	// TODO(knz): this should not use internalCompare any more.
 	return a.internalCompare(ctx, b)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DBool) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DBool)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	if !*d && *v {
+		return -1
+	}
+	if *d && !*v {
+		return 1
+	}
+	return 0
+}
+
+// internalCompare implements the Datum interface.
+func (d *DInt) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	var v DInt
+	switch t := UnwrapDatum(ctx, other).(type) {
+	case *DInt:
+		v = *t
+	case *DFloat, *DDecimal:
+		return -t.internalCompare(ctx, d)
+	default:
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	if *d < v {
+		return -1
+	}
+	if *d > v {
+		return 1
+	}
+	return 0
+}
+
+// internalCompare implements the Datum interface.
+func (d *DFloat) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	var v DFloat
+	switch t := UnwrapDatum(ctx, other).(type) {
+	case *DFloat:
+		v = *t
+	case *DInt:
+		v = DFloat(MustBeDInt(t))
+	case *DDecimal:
+		return -t.internalCompare(ctx, d)
+	default:
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	if *d < v {
+		return -1
+	}
+	if *d > v {
+		return 1
+	}
+	// NaN sorts before non-NaN (#10109).
+	if *d == v {
+		return 0
+	}
+	if math.IsNaN(float64(*d)) {
+		if math.IsNaN(float64(v)) {
+			return 0
+		}
+		return -1
+	}
+	return 1
+}
+
+// internalCompare implements the Datum interface.
+func (d *DDecimal) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v := ctx.getTmpDec()
+	switch t := UnwrapDatum(ctx, other).(type) {
+	case *DDecimal:
+		v = &t.Decimal
+	case *DInt:
+		v.SetCoefficient(int64(*t)).SetExponent(0)
+	case *DFloat:
+		if _, err := v.SetFloat64(float64(*t)); err != nil {
+			panic(err)
+		}
+	default:
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	// NaNs sort first in SQL.
+	if dn, vn := d.Form == apd.NaN, v.Form == apd.NaN; dn && !vn {
+		return -1
+	} else if !dn && vn {
+		return 1
+	} else if dn && vn {
+		return 0
+	}
+	return d.Cmp(v)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DString) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DString)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	if *d < *v {
+		return -1
+	}
+	if *d > *v {
+		return 1
+	}
+	return 0
+}
+
+// internalCompare implements the Datum interface.
+func (d *DCollatedString) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DCollatedString)
+	if !ok || d.Locale != v.Locale {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	return bytes.Compare(d.Key, v.Key)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DBytes) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DBytes)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	if *d < *v {
+		return -1
+	}
+	if *d > *v {
+		return 1
+	}
+	return 0
+}
+
+// internalCompare implements the Datum interface.
+func (d *DUuid) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DUuid)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	return bytes.Compare(d.GetBytes(), v.GetBytes())
+}
+
+// internalCompare implements the Datum interface.
+func (d *DIPAddr) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DIPAddr)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+
+	return d.IPAddr.Compare(&v.IPAddr)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DDate) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	var v DDate
+	switch t := UnwrapDatum(ctx, other).(type) {
+	case *DDate:
+		v = *t
+	case *DTimestamp, *DTimestampTZ:
+		return compareTimestamps(ctx, d, other)
+	default:
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	if *d < v {
+		return -1
+	}
+	if v < *d {
+		return 1
+	}
+	return 0
+}
+
+// internalCompare implements the Datum interface.
+func (d *DTime) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	return compareTimestamps(ctx, d, other)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DTimeTZ) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	return compareTimestamps(ctx, d, other)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DTimestamp) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	return compareTimestamps(ctx, d, other)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DTimestampTZ) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	return compareTimestamps(ctx, d, other)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DInterval) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DInterval)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	return d.Duration.Compare(v.Duration)
+}
+
+// internalCompare implements the Datum interface.
+func (d *DJSON) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DJSON)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	// No avenue for us to pass up this error here at the moment, but Compare
+	// only errors for invalid encoded data.
+	// TODO(justin): modify internalCompare to allow passing up errors.
+	c, err := d.JSON.Compare(v.JSON)
+	if err != nil {
+		panic(err)
+	}
+	return c
+}
+
+// internalCompare implements the Datum interface.
+func (d *DTuple) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DTuple)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	n := len(d.D)
+	if n > len(v.D) {
+		n = len(v.D)
+	}
+	for i := 0; i < n; i++ {
+		c := d.D[i].internalCompare(ctx, v.D[i])
+		if c != 0 {
+			return c
+		}
+	}
+	if len(d.D) < len(v.D) {
+		return -1
+	}
+	if len(d.D) > len(v.D) {
+		return 1
+	}
+	return 0
+}
+
+// internalCompare implements the Datum interface.
+func (d dNull) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		return 0
+	}
+	return -1
+}
+
+// internalCompare implements the Datum interface.
+func (d *DArray) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DArray)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	n := d.Len()
+	if n > v.Len() {
+		n = v.Len()
+	}
+	for i := 0; i < n; i++ {
+		c := d.Array[i].internalCompare(ctx, v.Array[i])
+		if c != 0 {
+			return c
+		}
+	}
+	if d.Len() < v.Len() {
+		return -1
+	}
+	if d.Len() > v.Len() {
+		return 1
+	}
+	return 0
+}
+
+// internalCompare implements the Datum interface.
+func (d *DOid) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	v, ok := UnwrapDatum(ctx, other).(*DOid)
+	if !ok {
+		panic(makeUnsupportedComparisonMessage(d, other))
+	}
+	if d.DInt < v.DInt {
+		return -1
+	}
+	if d.DInt > v.DInt {
+		return 1
+	}
+	return 0
+}
+
+// internalCompare implements the Datum interface.
+func (d *DOidWrapper) internalCompare(ctx *EvalContext, other Datum) int {
+	if other == DNull {
+		// NULL is less than any non-NULL value.
+		return 1
+	}
+	if v, ok := other.(*DOidWrapper); ok {
+		return d.Wrapped.internalCompare(ctx, v.Wrapped)
+	}
+	return d.Wrapped.internalCompare(ctx, other)
+}
+
+// internalCompare implements the Datum interface.
+func (d *Placeholder) internalCompare(ctx *EvalContext, other Datum) int {
+	return d.mustGetValue(ctx).internalCompare(ctx, other)
 }
